Rewrite statsdash
This is a complete rewrite of statsdash. So far, it just does what it
did before, but has some additional graph breakdowns.

I included the metrics-graphics stuff, because I plan to use that in
conjunction with the Input API to pull Input data to get a feel for
whether things are "normal" or not.